+++ packages/article-flag/article-flag.showcase.js
@@ -5,7 +5,8 @@ import {
   UpdatedArticleFlag,
   ExclusiveArticleFlag,
   SponsoredArticleFlag,
-  LongReadArticleFlag
+  LongReadArticleFlag,
+  LiveArticleFlag
 } from "./src/article-flag";
 
 export default {
@@ -13,6 +14,11 @@ export default {
     {
       component: () => <NewArticleFlag />,
       name: "Article Flag (New)",
+      type: "story"
+    },
+    {
+      component: () => <LiveArticleFlag />,
+      name: "Article Flag (Live)",
       type: "story"
     },
     {
@@ -44,6 +50,7 @@ export default {
       component: () => (
         <ArticleFlags
           flags={[
+            { expiryTime: "2030-03-13T12:00:00.000Z", type: "LIVE" },
             { expiryTime: "2030-03-13T12:00:00.000Z", type: "UPDATED" },
             { expiryTime: "2030-03-14T12:00:00.000Z", type: "EXCLUSIVE" },
             { expiryTime: "2030-03-14T12:00:00.000Z", type: "NEW" },
@@ -59,6 +66,7 @@ export default {
       component: () => (
         <ArticleFlags
           flags={[
+            { expiryTime: "2030-03-13T12:00:00.000Z", type: "LIVE" },
             { expiryTime: "2030-03-13T12:00:00.000Z", type: "UPDATED" },
             { expiryTime: "2030-03-14T12:00:00.000Z", type: "EXCLUSIVE" },
             { expiryTime: "2030-03-14T12:00:00.000Z", type: "NEW" },