+++ packages/article-magazine-standard/src/article-magazine-standard.web.js
@@ -68,6 +68,7 @@ class ArticlePage extends Component {
       commentingConfig,
       paidContentClassName,
       isPreview,
+      swgProductId,
       additionalRelatedArticlesFlag,
       inlineRelatedArticlesFlag,
       algoliaSearchKeys,
@@ -91,6 +92,7 @@ class ArticlePage extends Component {
         commentingConfig={commentingConfig}
         paidContentClassName={paidContentClassName}
         isPreview={isPreview}
+        swgProductId={swgProductId}
         additionalRelatedArticlesFlag={additionalRelatedArticlesFlag}
         inlineRelatedArticlesFlag={inlineRelatedArticlesFlag}
         algoliaSearchKeys={algoliaSearchKeys}