+++ packages/article-main-comment/src/article-main-comment.web.js
@@ -63,6 +63,7 @@ class ArticlePage extends Component {
       commentingConfig,
       paidContentClassName,
       isPreview,
+      swgProductId,
       additionalRelatedArticlesFlag,
       inlineRelatedArticlesFlag,
       algoliaSearchKeys,
@@ -86,6 +87,7 @@ class ArticlePage extends Component {
         commentingConfig={commentingConfig}
         paidContentClassName={paidContentClassName}
         isPreview={isPreview}
+        swgProductId={swgProductId}
         additionalRelatedArticlesFlag={additionalRelatedArticlesFlag}
         inlineRelatedArticlesFlag={inlineRelatedArticlesFlag}
         algoliaSearchKeys={algoliaSearchKeys}