+++ packages/article-main-standard/src/article-main-standard.web.js
@@ -83,6 +83,7 @@ class ArticlePage extends Component {
       commentingConfig,
       paidContentClassName,
       isPreview,
+      swgProductId,
       additionalRelatedArticlesFlag,
       inlineRelatedArticlesFlag,
       algoliaSearchKeys,
@@ -108,6 +109,7 @@ class ArticlePage extends Component {
           commentingConfig={commentingConfig}
           paidContentClassName={paidContentClassName}
           isPreview={isPreview}
+          swgProductId={swgProductId}
           additionalRelatedArticlesFlag={additionalRelatedArticlesFlag}
           inlineRelatedArticlesFlag={inlineRelatedArticlesFlag}
           algoliaSearchKeys={algoliaSearchKeys}