+++ packages/author-profile/CHANGELOG.md
@@ -3,6 +3,30 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [6.4.8](https://github.com/newsuk/times-components/compare/@times-components/author-profile@6.4.7...@times-components/author-profile@6.4.8) (2021-11-23)
+
+**Note:** Version bump only for package @times-components/author-profile
+
+
+
+
+
+## [6.4.7](https://github.com/newsuk/times-components/compare/@times-components/author-profile@6.4.6...@times-components/author-profile@6.4.7) (2021-11-23)
+
+**Note:** Version bump only for package @times-components/author-profile
+
+
+
+
+
+## [6.4.6](https://github.com/newsuk/times-components/compare/@times-components/author-profile@6.4.5...@times-components/author-profile@6.4.6) (2021-11-16)
+
+**Note:** Version bump only for package @times-components/author-profile
+
+
+
+
+
 ## [6.4.5](https://github.com/newsuk/times-components/compare/@times-components/author-profile@6.4.4...@times-components/author-profile@6.4.5) (2021-10-28)
 
 **Note:** Version bump only for package @times-components/author-profile