+++ packages/ssr/CHANGELOG.md
@@ -3,6 +3,41 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+# [2.28.0](https://github.com/newsuk/times-components/compare/@times-components/ssr@2.27.23...@times-components/ssr@2.28.0) (2021-11-24)
+
+
+### Features
+
+* **GEA-255:** Add swgProductId to structured data ([#2817](https://github.com/newsuk/times-components/issues/2817)) ([747323b](https://github.com/newsuk/times-components/commit/747323b115a99afa26022d714026ade45423ce9e))
+
+
+
+
+
+## [2.27.23](https://github.com/newsuk/times-components/compare/@times-components/ssr@2.27.22...@times-components/ssr@2.27.23) (2021-11-23)
+
+**Note:** Version bump only for package @times-components/ssr
+
+
+
+
+
+## [2.27.22](https://github.com/newsuk/times-components/compare/@times-components/ssr@2.27.21...@times-components/ssr@2.27.22) (2021-11-23)
+
+**Note:** Version bump only for package @times-components/ssr
+
+
+
+
+
+## [2.27.21](https://github.com/newsuk/times-components/compare/@times-components/ssr@2.27.20...@times-components/ssr@2.27.21) (2021-11-16)
+
+**Note:** Version bump only for package @times-components/ssr
+
+
+
+
+
 ## [2.27.20](https://github.com/newsuk/times-components/compare/@times-components/ssr@2.27.19...@times-components/ssr@2.27.20) (2021-10-28)
 
 **Note:** Version bump only for package @times-components/ssr